Rename /proc to /realproc temporarily
It seems that /proc tree is created automatically by docker runner, so
making this dir later is ignored and not committed. The solution is the
same as for /dev, so call it /realdev for the time is in container's
root dir. Later, when cpio packing is attempted, it can be renamed back
to /proc, as it is in subdir at this time.